+++ core/design-system/src/main/kotlin/com/team/bottles/core/designsystem/components/textfield/LineTextField.kt
@@ -227,17 +227,6 @@ fun BottlesLineTextFieldWithTrailingButton(
     }
 }
 
-@Composable
-fun BottlesTextFieldErrorCaption(
-    text: String
-) {
-    Text(
-        text = text,
-        style = BottlesTheme.typography.caption,
-        color = BottlesTheme.color.text.errorPrimary
-    )
-}
-
 /*==============Preview==============*/
 
 private data class TextFieldUiState(
