+++ core/design-system/src/main/kotlin/com/team/bottles/core/designsystem/components/textfield/TextFieldErrorCaption.kt
@@ -1,0 +1,16 @@
+package com.team.bottles.core.designsystem.components.textfield
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import com.team.bottles.core.designsystem.theme.BottlesTheme
+
+@Composable
+fun BottlesTextFieldErrorCaption(
+    text: String
+) {
+    Text(
+        text = text,
+        style = BottlesTheme.typography.caption,
+        color = BottlesTheme.color.text.errorPrimary
+    )
+}
